feat!: don't pass lhs to rhs
In previous, this plugin passes string to rhs to notify what key is used
to summon it. But this may cause some problems. For example, passing a
function with parameters to rhs may cause invalid arguments error.
I introduced this feature to change the behavior of rhs depende on the
pressed keys. But this can be achieved by separating it.
So I decided to remove this feature.