feat: add `set` and `del`, and deprecate `register`
As this plugin doesn't provide an api for deleting keymaps from specific
submode. So, we introduced a `vim.keymap.set` and `vim.keymap.del` like
interface: `submode.set` and `submode.del`.
These api is almost compatible with `vim.keymap.*` except `submode.del`
doesn't accept any options.
As long as `submode.register` can be replaced with `submode.set`, so we
deprecated this api and will remove it in version 2.0.0.