feat: add `submode.default` for define default mapping
The `submode.create` has ability to define default mappings to the
submode, but it makes code longer because it require normally 3 to 5
line for each definition.
So, this commit introduce a new api `submode.default`. This api is
compatible with `vim.keymap.set`, as `submode.create` accept similar
value for each maping.
With this api, user can define default mapping as little as 1 to 3 lines
for each mapping.