feat!: remove `default` and `seal`, and change signature of `submode.create`
As default mappings is expected not to be defined after the construction
of submode is done.
I introduced `seal` to prevent it, but it must be called explicitly, and
so that I also introduced a feature implicitly seal submode.
Now, there exists three way to define default mappings:
- Passing tables to `submode.create`
- Passing a callback to `submode.create`
- Call `submode.default` and `submode.seal` manually.
I think it's redundant and api should provide only one way for it.
So, in this commit, I choose second way as a standard method and remove
any other method as the second way can produce short code and can
restrict scope of mapping definition.
In other words, I removed `default` and `seal` api and restrict the way
to define default mappings to passing callback to `submode.create`.
Also, I think some times user don't want to define default mappings and
only need user mappings via `submode.set`, so I make the third argument
of `submode.create` as optional.